requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-telegram-bot==20.3
